+++ precompiles/relay-encoder/src/lib.rs
@@ -26,10 +26,7 @@ use frame_support::{
 	ensure,
 };
 use pallet_staking::RewardDestination;
-use precompile_utils::{
-	revert, succeed, Bytes, EvmData, EvmDataReader, EvmDataWriter, EvmResult, FunctionModifier,
-	PrecompileHandleExt, RuntimeHelper,
-};
+use precompile_utils::prelude::*;
 use sp_core::{H256, U256};
 use sp_runtime::AccountId32;
 use sp_runtime::Perbill;
@@ -64,7 +61,7 @@ pub trait StakeEncodeCall {
 	fn encode_call(call: AvailableStakeCalls) -> Vec<u8>;
 }
 
-#[precompile_utils::generate_function_selector]
+#[generate_function_selector]
 #[derive(Debug, PartialEq)]
 enum Action {
 	EncodeBond = "encode_bond(uint256,uint256,bytes)",
